% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_qdata.R
\name{gen_qdata}
\alias{gen_qdata}
\title{Generate questionnaire data
A function to generate questionnaire data with normally distributed item score with a mean of 0.}
\usage{
gen_qdata(n, k, likert = FALSE)
}
\arguments{
\item{n}{sample size}

\item{k}{vector number of items per questionnaire}

\item{likert}{continuous scores transformed to likert item scores}
}
\value{
data.frame with `n` rows and `nq * k` columns
}
\description{
Generate questionnaire data
A function to generate questionnaire data with normally distributed item score with a mean of 0.
}
