% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNAR.R
\name{MNAR}
\alias{MNAR}
\title{Generate MNAR (missign not at random) data}
\usage{
MNAR(x, alpha, pattern, f, g = 4, a = pattern)
}
\arguments{
\item{x}{data frame where missing observations should be generated in.}

\item{alpha}{proportion of cases that will get a missing data pattern}

\item{pattern}{a matrix with ncol=ncol(data), nrow=numer of missing data patterns; for each patter 0 indicates missing and 1 observed.}

\item{f}{frequency of each pattern}

\item{g}{the odds of the patterns to occur, the strength of the mechanism}
}
\value{
A data frame that contains missing observations
}
\description{
Generate MNAR (missign not at random) data
}
\examples{
library(MASS)
x <- mvrnorm(n=100,mu=c(0,0,0), Sigma=matrix(c(5,1,1,1,5,1,1,1,5),3,3))
alpha <- 0.25
pattern <- matrix(c(1,1,0,1,0,1),2,3, byrow=T)
f <- c(0.5,0.5)
g <- c(4,4)
MNAR(x,alpha,pattern,f,g)
}
