% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{perform}
\alias{perform}
\title{Perform}
\format{
A data frame with 142 rows and 3 variables:
\describe{
  \item{gender}{gender, Man Woman}
  \item{weight}{weight in kilograms}
  \item{performance}{score on performance test}
}
}
\source{
data-raw/R/perform.R
}
\usage{
perform
}
\description{
A dataset containing the performance scores for men and women given their 
weight.
}
\keyword{datasets}
