% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{weightloss}
\alias{weightloss}
\title{weightloss}
\format{
A data frame with 550 rows and 4 variables:
\describe{
  \item{group}{group, control or intervention}
  \item{id}{subject id}
  \item{month}{months of measurement}
  \item{weight}{weight in kilograms}
}
}
\source{
data-raw/R/weightloss.R
}
\usage{
weightloss
}
\description{
A dataset containing the weight data for a weight loss intervention study.
}
\keyword{datasets}
