% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ts.R
\name{calculate_ts}
\alias{calculate_ts}
\title{Calculate total scores}
\usage{
calculate_ts(x, k, .fun = sum)
}
\arguments{
\item{x}{data.frame with item scores that are to be summarized into a
total score.}

\item{k}{number of items per questionnaire}

\item{.fun}{function used to summarise the item scores}
}
\value{
data.frame with `ncol = nq` and `nrow = nrow(x)`.
}
\description{
Calculate total scores
}
\examples{
x <- gen_qdata(n=100, k= c(10,10,10,10,10))
calculate_ts(x, c(10,10,10,10,10), sum)
calculate_ts(x, c(10,10,10,10,10), mean)
}
