% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ts.R
\name{calculate_ts}
\alias{calculate_ts}
\title{Calculate total scores}
\usage{
calculate_ts(x, nq, k)
}
\arguments{
\item{x}{data.frame}

\item{nq}{number of questionnaires}

\item{k}{number of items per questionnaire}
}
\description{
Calculate total scores
}
