% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNAR.R
\name{MNAR}
\alias{MNAR}
\title{Generate MNAR (missign not at random) data}
\usage{
MNAR(
  x,
  alpha,
  pattern = "random",
  f = rep(1/nrow(pattern)),
  g = 4,
  npattern = NULL,
  a = pattern
)
}
\arguments{
\item{x}{data frame where missing observations should be generated in.}

\item{alpha}{proportion of cases that will get a missing data pattern}

\item{pattern}{a matrix with ncol=ncol(data), nrow=number of missing data 
patterns; for each patter 0 indicates missing and 1 observed.
`pattern = "random"` will generate `npattern` random patterns.}

\item{f}{frequency of each pattern}

\item{g}{the odds of the patterns to occur, the strength of the mechanism}

\item{npattern}{the number of patterns when patterns are randomly generated.}

\item{a}{linear predictor for missings}
}
\value{
A data frame that contains missing observations
}
\description{
Generate MNAR (missign not at random) data
}
\examples{
x <- MASS::mvrnorm(n=100,mu=c(0,0,0), Sigma=matrix(c(5,1,1,1,5,1,1,1,5),3,3))
alpha <- 0.25
pattern <- matrix(c(1,1,0,1,0,1),2,3, byrow=TRUE)
f <- c(0.5,0.5)
g <- c(4,4)
MNAR(x,alpha,pattern,f,g)
}
